Fin de journée 21.05.2021

+++ PasswordVault/PasswordVault/UsrCtrlAccount.cs
@@ -1,10 +1,4 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.ComponentModel;
-using System.Data;
-using System.Drawing;
-using System.Linq;
-using System.Text;
 using System.Threading.Tasks;
 using System.Windows.Forms;
 
@@ -12,9 +6,14 @@ namespace PasswordVault
 {
     public partial class UsrCtrlAccount : UserControl
     {
-        FormMain formMain;
-        Account account;
+        FormMain formMain; // Property
+        Account account; // Property
 
+        /// <summary>
+        /// Constructor
+        /// </summary>
+        /// <param name="formMain">Instance of the main form</param>
+        /// <param name="account">Account to display</param>
         public UsrCtrlAccount(FormMain formMain, Account account)
         {
             InitializeComponent();
@@ -22,34 +21,68 @@ namespace PasswordVault
             this.account = account;
         }
 
+        /// <summary>
+        /// Load method of the form
+        /// </summary>
+        /// <param name="sender"></param>
+        /// <param name="e"></param>
         private void UsrCtrlAccount_Load(object sender, EventArgs e)
         {
             lblTitle.Text = account.Title;
         }
 
+        /// <summary>
+        /// Opens the modify form
+        /// </summary>
+        /// <param name="sender"></param>
+        /// <param name="e"></param>
         private void btnModify_Click(object sender, EventArgs e)
         {
             FormModify formModify = new FormModify(formMain, account);
             formModify.ShowDialog();
         }
 
+        /// <summary>
+        /// Deletes the account and its history
+        /// </summary>
+        /// <param name="sender"></param>
+        /// <param name="e"></param>
         private void btnDelete_Click(object sender, EventArgs e)
         {
-            SQLiteDataAccess.DeleteAccount(account);
-            formMain.ReloadAccountList(SQLiteDataAccess.LoadAccounts());
+            // Asks for a confirmation
+            var confirmResult = MessageBox.Show("Are you sure to delete this account ??", "Confirm Delete!!", MessageBoxButtons.YesNo);
+            if (confirmResult == DialogResult.Yes)
+            {
+                // Deletes account and history
+                SQLiteDataAccess.DeleteAccount(account);
+                formMain.ReloadAccountList(SQLiteDataAccess.LoadAccounts());
+            }
         }
 
+        /// <summary>
+        /// Copies the username to clipboard
+        /// </summary>
+        /// <param name="sender"></param>
+        /// <param name="e"></param>
         private void btnCopyUsername_Click(object sender, EventArgs e)
         {
             Clipboard.SetText(account.Login);
         }
 
+        /// <summary>
+        /// Copies the password to clipboard and removes it after 15 seconds
+        /// </summary>
+        /// <param name="sender"></param>
+        /// <param name="e"></param>
         private void btnCopyPassword_Click(object sender, EventArgs e)
         {
             Clipboard.SetText(account.Password);
             ClearClipboard();
         }
 
+        /// <summary>
+        /// Clears the clipboard after 15 seconds
+        /// </summary>
         private async void ClearClipboard()
         {
             Task.Delay(15000).Wait();
